kinematics/code/requirements.txt
matplotlib
numpy
scipy.io

